+++ NEWS.md
@@ -3,6 +3,7 @@
 * `nflseedR_compute_results()` correctly adjusts Elo difference for postseason games. (#54)
 * `nfl_standings()` returns `exit` value consistent with `nfl_simulations()` if argument `ranks = "DRAFT"`. (#56)
 * `nfl_standings_prettify()` shows `exit` (if available). (#56)
+* `nfl_standings()` now warns the user when the `games` argument includes both `"season"` and `"sim"` to avoid confusion. (#58)
 
 # nflseedR 2.0.0